fix: openid4vp, test import file mismatch, move test in correct module (view: 444#discussion_r2171249896, #066dc830)
